--- client/src/features/question-composer/create/QuestionComposer.tsx
@@ -1,57 +1,0 @@
-import styled from 'styled-components';
-
-import { boardType, boardMutation } from '@/src/entities/qnaboard';
-
-import {
-  Container,
-  ErrorBoundary,
-  ErrorFallback,
-  Spinner,
-} from '@/src/shared/ui';
-import { styleMixin, V } from '@/src/shared/styles';
-import { usePageRouter } from '@/src/shared/hooks';
-
-import { Suspense } from 'react';
-import QuestionCreateForm from './QuestionCreateForm';
-
-function QuestionComposer() {
-  const { moveBoardMainPage } = usePageRouter();
-  const { mutate: writeBoard } = boardMutation.useWriteBoard();
-
-  const onSubmit = (data: boardType.IWriteQnaBoard) => {
-    writeBoard(data);
-    moveBoardMainPage();
-  };
-
-  return (
-    <ComposerContainer>
-      <ErrorBoundary
-        renderFallback={(_, onReset) => (
-          <ErrorFallback
-            title="데이터를 불러올 수 없습니다."
-            message={
-              '자소서 데이터를 불러오는 과정에서 문제가 발생했습니다. \n 잠시 후 다시 시도해 주세요.'
-            }
-            onReset={onReset}
-          />
-        )}
-      >
-        <Suspense fallback={<Spinner />}>
-          <QuestionCreateForm onSubmit={onSubmit} />
-        </Suspense>
-      </ErrorBoundary>
-    </ComposerContainer>
-  );
-}
-
-export default QuestionComposer;
-
-const ComposerContainer = styled(Container.ItemBase)`
-  ${styleMixin.Flex('flex-start')}
-  width: 100%;
-  height: 100%;
-  gap: 30px;
-  @media screen and (max-width: ${V.mediaWeb}) {
-    ${styleMixin.Column('flex-start')}
-  }
-`;
