+++ client/src/features/question-composer/create/ui/coverletter-preview/CoverLetterDetail.tsx
