+++ client/src/features/question-composer/create/ui/coverletter-preview/index.ts
@@ -1,0 +1,1 @@
+export { default as CoverLetterDetail } from './CoverLetterDetail';
