+++ client/src/features/question-composer/create/ui/coverletter-selector/index.ts
@@ -1,0 +1,1 @@
+export { default as CoverLetterSelector } from './CoverLetterSelector';
