+++ client/src/features/question-composer/create/ui/index.ts
@@ -1,0 +1,1 @@
+export { default as QuestionCreateForm } from './QuestionCreateForm';
