+++ client/src/features/question-composer/index.ts
@@ -1,2 +1,2 @@
-export { default as QuestionComposer } from './create/QuestionComposer';
+export { QuestionCreateForm } from './create/ui';
 export { default as QuestionEditComposer } from './edit/QuestionEditComposer';
